Update to .NET 8, misc refresh

+++ RazorPagesPizza/Pages/Pizza.cshtml.cs
@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using RazorPagesPizza.Models;
@@ -5,8 +6,10 @@ using RazorPagesPizza.Services;
 
 namespace RazorPagesPizza.Pages
 {
+    [Authorize]
     public class PizzaModel : PageModel
     {
+        public bool IsAdmin => HttpContext.User.HasClaim("IsAdmin", bool.TrueString);
         public List<Pizza> pizzas = new();
 
         [BindProperty]
@@ -19,6 +22,7 @@ namespace RazorPagesPizza.Pages
 
         public IActionResult OnPost()
         {
+            if (!IsAdmin) return Forbid();
             if (!ModelState.IsValid)
             {
                 return Page();
@@ -29,6 +33,7 @@ namespace RazorPagesPizza.Pages
 
         public IActionResult OnPostDelete(int id)
         {
+            if (!IsAdmin) return Forbid();
             PizzaService.Delete(id);
             return RedirectToAction("Get");
         }
